[Kernel/Multitasking] Overhauled the entire multitasking subsystem.
Now using a simplified version of the O(1) scheduler with less
context switching overhead, as well as less time spent running
idle CPU tasks, which are now only ran as fall-back in case no
other tasks are available. Many other improvements for things like
Thread/Process/Sleep.

+++ Kernel/Core/Syscall.h
@@ -9,7 +9,7 @@ class Syscall {
     MAKE_STATIC(Syscall);
 
 public:
-    static void exit(u8 exit_code);
+    static void exit(size_t exit_code);
     static void debug_log(const char* string);
 };
 }
